fix section unrender, all render tests passing
