+++ src/viewmodel/models/ArrayMemberReference.js
@@ -1,31 +1,7 @@
-import Reference from './Reference';
+// import Reference from './Reference';
 
-class ArrayMemberReference extends Reference {
+// import { Reference } from './Context';
 
-	constructor ( index ) {
-		super( '' + index );
-		this.index = index;
-	}
 
-	resolve () {
-		if ( this.resolved ) {
-			return;
-		}
 
-		if ( !this.parent.members ) {
-			this.parent.getOrCreateMembers();
-		}
-
-		let resolved;
-
-		if ( resolved = this.resolved = this.parent.members[ this.index ] ) {
-			resolved.register( this, 'computed' );
-		}
-	}
-
-	set ( value ) {
-		this.parent.setMember( this.index, value );
-	}
-}
-
-export default ArrayMemberReference;
+// export default ArrayMemberReference;
