+++ src/viewmodel/models/Context.js
@@ -2,10 +2,9 @@ import { addToArray, removeFromArray } from 'utils/array';
 import { isArray, isObject, isNumeric } from 'utils/is';
 import createBranch from 'utils/createBranch';
 
-import Reference from './Reference';
-import ArrayMemberReference from './ArrayMemberReference';
-import HashPropertyReference from './ArrayMemberReference';
-import { IndexSpecial, KeySpecial, KeypathSpecial } from './Specials';
+// import ArrayMemberReference from './ArrayMemberReference';
+// import HashPropertyReference from './ArrayMemberReference';
+// import { IndexSpecial, KeySpecial, KeypathSpecial } from './Specials';
 
 import PropertyStore from '../stores/PropertyStore';
 import StateStore from '../stores/StateStore';
@@ -777,6 +776,251 @@ Context.prototype = {
 
 }
 
+const noopStore = {};
+
+class IndexSpecial extends Context {
+
+	constructor () {
+		// babel bug
+		this.that = 0;
+		super( '@index', noopStore );
+	}
+
+	get () {
+		return this.parent.index;
+	}
+
+	set () {
+		throw new Error('cannot set @index');
+	}
+
+	// required as child or Reference
+	reset () {
+
+	}
+}
+
+class KeySpecial extends Context {
+
+	constructor () {
+		// babel bug
+		this.that = 0;
+		super( '@key', noopStore );
+	}
+
+	get () {
+		return this.parent.getKey();
+	}
+
+	set () {
+		throw new Error('cannot set @key');
+	}
+
+	// required as child of Reference
+	reset () {
+
+	}
+}
+
+class KeypathSpecial extends Context {
+
+	constructor () {
+		// babel bug
+		this.that = 0;
+		super( '@keypath', noopStore );
+	}
+
+	get () {
+		return this.parent.getKeypath();
+	}
+
+	set () {
+		throw new Error('cannot set @keypath');
+	}
+}
+
+
+class Reference extends Context {
+
+	constructor ( key ) {
+		this.resolved = null;
+		super( key, noopStore );
+	}
+
+	getJoinKey () {
+		return this.key;
+	}
+
+	get () {
+		this.resolve();
+
+		let resolved;
+		if ( resolved = this.resolved ) {
+			return resolved.get();
+		}
+	}
+
+	getSettable () {
+		this.resolve();
+
+		let resolved;
+		if ( resolved = this.resolved ) {
+			return resolved.getSettable();
+		}
+		else {
+			// TODO ???
+			throw new Error('ArrayMemberReference not settable, need to force?')
+		}
+	}
+
+	resolve () {
+		if ( this.resolved ) {
+			return;
+		}
+
+		let resolved,
+			joinParent = this.parent.getJoinModel(),
+			key = this.getJoinKey();
+
+		if ( joinParent && key != null ) {
+			resolved = this.resolved = joinParent.join( key );
+		}
+
+		if ( resolved ) {
+			resolved.register( this, 'computed' );
+		}
+	}
+
+	// Don't know if this is answer to not re-resolving with old value
+	// on cascade. Probably a better option...
+	cascadeDown () {
+		// this.createOrReconcileMembers( this.get() );
+		this.cascadeChildren( this.members );
+		this.cascadeChildren( this.properties );
+	}
+
+
+	set ( value ) {
+		this.resolve();
+
+		let resolved = this.resolved;
+		if ( !resolved ) {
+			// TODO force resolve?
+			if ( typeof value !== 'undefined' ) {
+				throw new Error('Reference set called without resolved.');
+			}
+			return;
+		}
+
+		resolved.set( value );
+	}
+
+	reset () {
+		if ( this.resolved ) {
+			this.resolved.unregister( this, 'computed' );
+			this.resolved = null;
+		}
+
+		this.mark();
+
+		this.resetChildren( this.properties );
+		// TODO: do members actually need to be reset ???
+		this.resetChildren( this.members );
+
+	}
+
+	resetChildren ( children ) {
+		if ( !children ) { return; }
+
+		for( let i = 0, l = children.length; i < l; i++ ) {
+			children[i].reset();
+		}
+	}
+
+	getJoinModel () {
+		this.resolve();
+		let resolved = this.resolved;
+		if ( !resolved ) {
+			// TODO:  create new Unresolved() ????
+			throw new Error('Reference getJoinModel called without resolved.');
+		}
+		return resolved;
+	}
+
+	createChild ( key ) {
+		return new Reference( key );
+	}
+
+	doJoin ( keys, testFirstKey, firstKey ) {
+
+		this.resolve();
+
+		if ( !this.resolved ) {
+			throw new Error('attempt to join unresolved reference');
+		}
+
+		return super.doJoin( keys, testFirstKey, firstKey );
+	}
+}
+
+class HashPropertyReference extends Reference {
+
+	constructor ( key, index ) {
+		// babel bug
+		this.that = 0;
+		super( key );
+		this.index = index;
+	}
+
+	resolve () {
+		if ( this.resolved ) {
+			return;
+		}
+
+		if ( !this.parent.members ) {
+			this.parent.getOrCreateMembers();
+		}
+
+		let resolved;
+
+		if ( resolved = this.resolved = this.parent.tryJoin( this.key ) ) {
+			resolved.register( this, 'computed' );
+		}
+	}
+
+	// Don't think this is need for Hash Member
+	// set ( value ) {
+	// 	this.parent.setMember( this.index, value );
+	// }
+}
+
+class ArrayMemberReference extends Reference {
+
+	constructor ( index ) {
+		super( '' + index );
+		this.index = index;
+	}
+
+	resolve () {
+		if ( this.resolved ) {
+			return;
+		}
+
+		if ( !this.parent.members ) {
+			this.parent.getOrCreateMembers();
+		}
+
+		let resolved;
+
+		if ( resolved = this.resolved = this.parent.members[ this.index ] ) {
+			resolved.register( this, 'computed' );
+		}
+	}
+
+	set ( value ) {
+		this.parent.setMember( this.index, value );
+	}
+}
 
 function hasChildFor ( value, key ) {
 	if ( value == null ) {
@@ -789,3 +1033,4 @@ function hasChildFor ( value, key ) {
 }
 
 export default Context;
+export { Reference };
