+++ src/viewmodel/models/HashPropertyReference.js
@@ -1,32 +1,9 @@
-import Reference from './Reference';
+/* moved to Context */
 
-class HashPropertyReference extends Reference {
+// import Reference from './Reference';
 
-	constructor ( key, index ) {
-		// babel bug
-		this.that = 0;
-		super( key );
-		this.index = index;
-	}
+// import { Reference } from './Context';
 
-	resolve () {
-		if ( this.resolved ) {
-			return;
-		}
 
-		if ( !this.parent.members ) {
-			this.parent.getOrCreateMembers();
-		}
 
-		let resolved;
-
-		if ( resolved = this.resolved = this.parent.tryJoin( this.key ) ) {
-			resolved.register( this, 'computed' );
-		}
-	}
-
-	// Don't think this is need for Hash Member
-	// set ( value ) {
-	// 	this.parent.setMember( this.index, value );
-	// }
-}
+// export default HashPropertyReference;
