+++ src/virtualdom/items/Section/prototype/unbind.js
@@ -4,8 +4,7 @@ import unbind from '../../shared/unbind';
 
 export default function Section$unbind () {
 	this.fragments.forEach( unbindFragment );
-	this.fragmentsToRender.forEach( f => removeFromArray( this.fragments, f ) );
-	this.fragmentsToRender = [];
+
 	unbind.call( this );
 
 	this.length = 0;
